+++ src/GoldyBot/database/member/money.py
@@ -16,24 +16,28 @@ class Money():
         except KeyError:
             return currency.default_bal
 
-    async def give_money(self, currency_class:Currency, amount:int):
+    async def give_money(self, currency_class:Currency, amount:int) -> bool:
         """Gives the member the specified amount of money on that specified currency."""
         currency = currency_class
 
         # Add to previous bal in database.
-        await self.actual_member.database.edit("global", {"_id":int(self.actual_member.member_id)}, {
+        result = await self.actual_member.database.edit("global", {"_id":int(self.actual_member.member_id)}, {
             self.actual_member.member_id: {
                 currency.code_name:  (await self.get_money(currency) + amount)
             }
         })
 
-    async def take_money(self, currency_class:Currency, amount:int):
+        return result
+
+    async def take_money(self, currency_class:Currency, amount:int) -> bool:
         """Takes from the member a specified amount of money on that specified currency."""
         currency = currency_class
 
         # Take from previous bal in database.
-        await self.actual_member.database.edit("global", {"_id":int(self.actual_member.member_id)}, {
+        result = await self.actual_member.database.edit("global", {"_id":int(self.actual_member.member_id)}, {
             self.actual_member.member_id: {
                 currency.code_name:  (await self.get_money(currency) - amount)
             }
         })
+
+        return result