+++ src/CHANGELOG.rst
@@ -1,7 +1,21 @@
 Change Log
 ==========
 
-v4.0dev23 = BREAKING CHANGE! ()
+v4.0dev24 = New Dev Release (???)
+---------------------------
+- Better typing added for "slash_options".
+- give_money() and take_money() now returns boolean.
+- Errors are now accessible right from GoldyBot's main module.
+- Improved error handler in /timestamp command.
+- Better typing in goldy bot utility module.
+- Added Goldy Bot hearts class.
+- Removed dev command /stop, you can now only stop the bot from the console.
+- Improved Goldy Bot stop method. (it functions much better now)
+- Removed all lower case 'Colours' variable in main GoldyBot module. ('colours' is now typed as 'Colours') (Backwards compatibility was added so this isn't a breaking change for extensions.)
+- Added most common colour in image picker method in the GoldyBot.Colours class.
+- Added the WebFile() object.
+
+v4.0dev23 = BREAKING CHANGE! (16/09/2022)
 ---------------------------
 - Optimizations/speed ups.
 - Role object parameters have been changed, so make sure you update it in your code. A complete new redesign has been done to it. (BREAKING CHANGE!)
@@ -20,7 +34,7 @@ v4.0dev10 = New Release (25/06/2022)
 
 v4.0dev9 - BREAKING CHANGE! (26/04/2022)
 ---------------------------
-- 'toggle_slash_cmd' and 'toggle_normal_cmd' have been changed to 'slash_cmd_only' and 'normal_cmd_only'. I belive this is much easier to understand.
+- 'toggle_slash_cmd' and 'toggle_normal_cmd' have been changed to 'slash_cmd_only' and 'normal_cmd_only'. I believe this is much easier to understand.
 
 v4.0dev1 - First Dev Release (15/04/2022)
 ---------------------------